Improve formatting to match core-library style
* This commit removes the newlines printed between each declaration
  and only prints newlines before certain top-level declarations.

For instance, imports get no newline but most everything else does.
The one exception are function/value declarations because they
should not get a newline if they are not a top-level declaration.
(However, they should get a newline if they are!)

* Module declarations should now be formatted like they are in the
  core libraries (excluding indentation width.) Specifically:

  * No blank line should appear before the declaration if comments
    are absent.
  * Only one space should appear between the module name and `where`
    keyword if the export list is missing.

Export lists should now be formatted like this:

```purescript
module Data.Bounded
    ( class Bounded
    , bottom
    , top
    , module Data.Ord
    ) where
```

* Ensure files end with a new line
* Default to two-space indentation

Hopefully this isn't controversial: the core libraries use 2-space
indents exclusively. (I personally prefer four spaces, but it's
mostly arbitrary.)

* Don't insert space before (..) in import/export lists
* Don't place a newline after an equals sign

This aligns the formatting more closely with core-library style.
`where` clauses are now indented to match how it's done in the
core libraries as well.

* Fix indentation of multiline functions

This ensures that multiline functions are indented. `case`
expressions are also brought down one line from a function
definition to match core-library style. Do blocks are
over-indented, however.

* Change default width to 80 characters

Confirmed this with paf31.

* Wrap long lines

This enables wrapping of long types, if expressions, function/value
declarations, and case expressions.

* Eliminate space after constructors with no parameters
* Eliminate parentheses around lone constraints

This means that `(Show a) =>` will become `Show a =>`.

* Pretty-prints the comments associated with `PositionedValue`s
  and pretty-prints `PositionedDoNotationElement`s.
* Fix pretty-printing of empty data declarations
* Pretty-print positioned binders and typed binders