Add options
This adds the following options to psc-format:
* Indentation width (--indent)
* Maximum page width (--width)
* Use Unicode characters (--unicode)

Additionally, --input was removed and is now the first argument
instead.

Note that this was a fairly large change. In order to pass the
options through to the pretty print functions they had to be removed
from the `Pretty` type class instances. Additionally,
Binder, Expr, and Literal had recursive definitions so they ended up
being placed in a single module for the time being.